+++ lib/view/page/mypage/my_page.dart
@@ -1,7 +1,11 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:iris_flutter/view/comm/custom_appbar.dart';
+import 'package:iris_flutter/view/page/mypage/bookmark_info.dart';
+import 'package:iris_flutter/view/page/mypage/written_info.dart';
+import 'package:iris_flutter/view/page/mypage/written_reports.dart';
 
 class MyPage extends StatefulWidget {
   const MyPage({super.key});
@@ -62,7 +66,7 @@ class _MyPageState extends State<MyPage> {
                           topLeft: Radius.circular(15),
                           topRight: Radius.circular(15))),
                   onTap: () {
-                    print("작성한 실종 정보");
+                    Get.to(() => const WrittenInfo());
                   },
                   leading: Icon(
                     Icons.people,
@@ -75,7 +79,7 @@ class _MyPageState extends State<MyPage> {
                 ListTile(
                   tileColor: Theme.of(context).colorScheme.surfaceVariant,
                   onTap: () {
-                    print("작성한 제보 댓글");
+                    Get.to(() => const WrittenReports());
                   },
                   leading: const Icon(Icons.chat),
                   title: const Text("작성한 제보 댓글"),
@@ -89,7 +93,7 @@ class _MyPageState extends State<MyPage> {
                           bottomLeft: Radius.circular(15),
                           bottomRight: Radius.circular(15))),
                   onTap: () {
-                    print("북마크한 실종 정보");
+                    Get.to(() => const BookmarkInfo());
                   },
                   leading: const Icon(Icons.bookmark),
                   title: const Text("북마크한 실종 정보"),
